% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structure.R
\name{apply_structure_x}
\alias{apply_structure_x}
\title{Applies expected data structure for x variable and provides params list}
\usage{
apply_structure_x(x, data = FALSE, ..., strctr = FALSE)
}
\arguments{
\item{x}{The x-variable from the original call, passed through \code{syscalls}.}

\item{data}{An optional data frame including \code{x} column.}

\item{...}{Other arguments to the function (i.e., paired, alternative, etc.)}

\item{strctr}{Character vector, how to format data \code{c("numeric", "factor")}.}
}
\value{
Returns a params list
}
\description{
Applies expected data structure for x variable and provides params list
}
