% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-params.R
\name{at_get_fn_params}
\alias{at_get_fn_params}
\title{at_get_fn_params}
\usage{
at_get_fn_params(fn_name = NULL, pkg_name = NULL)
}
\arguments{
\item{fn_name}{The function for which parameters are to be extracted, either
as a simply function name along with 'package' parameter, or in double-colon
form as 'fn_name::pkg_name', in which case the 'pkg_name' parameter need not
be provided.}

\item{pkg_name}{The package defining the named function.}
}
\description{
Get parameters of a specified function
}
