% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-yaml.R
\name{at_yaml_template}
\alias{at_yaml_template}
\title{at_yaml_template}
\usage{
at_yaml_template(loc = tempdir())
}
\arguments{
\item{loc}{Location to generate template file. Append with filename and
'.yaml' suffix to overwrite default name of 'autotest.yaml', otherwise this
parameter will be used to specify directory only.}
}
\description{
Generate a 'yaml' template for an 'autotest'.
}
\seealso{
Other yaml:
\code{\link[=examples_to_yaml]{examples_to_yaml()}}
}
\concept{yaml}
