% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoClustR_v2AS_refactor.R
\name{AutoClustR}
\alias{AutoClustR}
\title{AutoClustR}
\usage{
AutoClustR(
  object,
  method = "CH",
  standardize = TRUE,
  k.param.space = c(10, 240, 10),
  resolution.space = c(0.1, 2.4, 0.1),
  n.iterations = 24,
  pc.use = 200,
  pc.estimate = 10,
  ...
)
}
\arguments{
\item{object}{The Seurat object to be tested}

\item{method}{Specify method used for cluster ranking. "CH" for Calisnki-Harabasz,
"DB" for Davies-Bouldin, "Sil" for silhouette. Default is "CH"}

\item{standardize}{Determines whether or not to standardize principal components. Default is TRUE}

\item{k.param.space}{Numeric vector of length three, specifying minimum, maximum, and step size. Default is c(10, 240, 10)}

\item{resolution.space}{Numeric vector of length three, specifyin gminimum, maximum, and step size. Default is c(0.1, 2.4, 0.1)}

\item{n.iterations}{Number of restarts for hill-climbing optimization.
Lower values will give lower runtimes but may not find optimal parameter values. Default is 24}

\item{pc.use}{Number of principal components used for segmented linear regression testing. Default is 200}

\item{pc.estimate}{Used by segemented function to estimate segmented linear regression breakpoint. Default is 10.}
}
\value{
A list containing a seurat object clustered using optimal parameters, a double containing the number of PC's used,
two matrices containing the number of clusters and the ranking for tested parameter combinations,
an elbow plot showing the segmented linear regression, and a heatmap showing ranknigs for tested parameter combinations
}
\description{
Determines the optimal number of principal components and optimal values for the resolution and K.param arguments used for clustering
}
\examples{
#Returns list with clustered seurat object and descriptive data
clustering_data <- AutoClustR(seurat.object, method = "CH", k.param.space = c(10, 120, 10))
}
