% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotest-package.R
\docType{package}
\name{autotest-package}
\alias{autotest}
\alias{autotest-package}
\title{autotest: Automatic Package Testing}
\description{
Automatic testing of R packages via a simple YAML schema.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/autotest/}
  \item \url{https://github.com/ropensci-review-tools/autotest}
  \item Report bugs at \url{https://github.com/ropensci-review-tools/autotest/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})

Authors:
\itemize{
  \item Mark Padgham \email{mark.padgham@email.com} (\href{https://orcid.org/0000-0003-2172-5265}{ORCID})
}

Other contributors:
\itemize{
  \item Jouni Helske (\href{https://orcid.org/0000-0001-7130-793X}{ORCID}) [contributor]
}

}
\keyword{internal}
