% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotest.R
\name{autotest}
\alias{autotest}
\title{autotest}
\usage{
autotest(yaml = NULL, filename = NULL, quiet = FALSE)
}
\arguments{
\item{yaml}{A 'yaml' template as a character vector, either hand-coded or
potentially loaded via \link{readLines} function or similar. Should generally
be left at default of 'NULL', with template specified by 'filename'
parameter.}

\item{filename}{Name (potentially including path) of file containing 'yaml'
template. See \link{at_yaml_template} for details of template. Default uses
template generated by that function, and held in local './tests' directory.}

\item{quiet}{If 'FALSE', provide printed output on screen.}
}
\value{
Nothing
}
\description{
Automatically test inputs to functions specified in a 'yaml' template.
}
\examples{
\dontrun{
exs <- examples_to_yaml (package = "stats", functions = "reshape")
res <- autotest (yaml = exs)
}
}
