% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotest-class.R
\name{autotest_obj}
\alias{autotest_obj}
\title{autotest_obj class definition}
\usage{
autotest_obj(
  package = NA_character_,
  package_loc = NULL,
  test_name = NA_character_,
  fn_name = NA_character_,
  parameters = list(),
  parameter_types = NA_character_,
  class = NULL,
  classes = NULL,
  env = new.env(),
  test = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{package}{Name of package for which object is to be constructed.}

\item{package_loc}{Location of package on local system (for source packages
only)}

\item{test_name}{Name of test (use \link{autotest_types} to get all test
names).}

\item{fn_name}{Name of function to be tested.}

\item{parameters}{Names of all parameters for that function.}

\item{parameter_types}{Types of input parameters.}

\item{class}{Class of an individual parameter.}

\item{classes}{Classes of all parameters.}

\item{env}{Environment in which tests are to be run.}

\item{test}{If \code{FALSE}, return only descriptions of tests which would be run
with \code{test = TRUE}, without actually running them.}

\item{quiet}{If \code{FALSE}, issue progress and other messages during testing of
object.}
}
\value{
An object of class \code{autotest_obj}, a list holding the parameters
passed as arguments, used internally to represent one autotest test case.
}
\description{
This function exists only to provide the class definitions for test objects,
and is not intended to be called directly.
}
\examples{
x <- autotest_obj (
    package = "stats",
    fn_name = "sd",
    test_name = "test"
)
}
\concept{class}
