% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-list.R
\name{autotest_results}
\alias{autotest_results}
\title{Create a `autotest_results` object from the test results
as stored in the ListReporter results field.}
\usage{
autotest_results(results)
}
\arguments{
\item{results}{a list as stored in ListReporter}
}
\value{
its list argument as a `autotest_results` object
}
\description{
Create a `autotest_results` object from the test results
as stored in the ListReporter results field.
}
\seealso{
ListReporter
}

