% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structure.R
\name{build_var}
\alias{build_var}
\title{Builds the var for data with and without data frame}
\usage{
build_var(var, strctr, data = FALSE)
}
\arguments{
\item{var}{The \code{v} from \code{apply_structure_}-functions.}

\item{strctr}{String of required structure.}

\item{data}{An optional data frame including the \code{v} column.}
}
\value{
Returns structure-formatted data.
}
\description{
Builds the var for data with and without data frame
}
\examples{
build_var(ToothGrowth$len, "numeric")
build_var(ToothGrowth["len"], "numeric")
build_var(ToothGrowth[["len"]], "numeric")
build_var("len", "numeric", ToothGrowth)
}
