% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structure.R
\name{build_var_name}
\alias{build_var_name}
\title{Builds the var_name for data without data frame}
\usage{
build_var_name(var_name)
}
\arguments{
\item{var_name}{The \code{v_name} from \code{apply_structure_}-functions.}
}
\value{
Returns a string with the formatted \code{v_var_name}.
}
\description{
Builds the var_name for data without data frame
}
