% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calc_space}
\alias{calc_space}
\title{Calculates amount of spaces for reporting}
\usage{
calc_space(x, params = NULL, min = 0)
}
\arguments{
\item{x}{The string for which spaces are calculated.}

\item{params}{vector of other strings in table.}

\item{min}{The minimum space distance (optional).}
}
\value{
Returns an integer of required space.
}
\description{
Calculates amount of spaces for reporting
}
\examples{
calc_space("This is a test", c("This", "is", "a", "test"), 34)
calc_space("This", c("This is a test", "is", "a", "test"), 34)
calc_space("This", c("This is a test", "is", "a", "test"))
}
