% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_normality.R
\name{check_normality}
\alias{check_normality}
\title{Performs a check on normality}
\usage{
check_normality(x, data = "", alpha = 0.05)
}
\arguments{
\item{x}{A numeric vector of data values.}

\item{alpha}{Maximum accepted p-value for significance tests.}
}
\value{
Returns a list of all results.
}
\description{
Performs a check on normality
}
\examples{
check_normality(ToothGrowth$len)
check_normality("len", ToothGrowth)
check_normality(ToothGrowth[["len"]])
check_normality(runif(233))
check_normality(rnorm(233))
check_normality("ERROR")
}
\seealso{
NCmisc::list.functions.in.file(filename = rstudioapi::getSourceEditorContext()$path)
}
