% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_normality.R
\name{check_normality.report}
\alias{check_normality.report}
\title{Class to build a full report for check_normality}
\usage{
check_normality.report(object)
}
\arguments{
\item{object}{Object of check_normality function}
}
\value{
Returns a full test report with simple figures
}
\description{
Class to build a full report for check_normality
}
\examples{
report(check_normality(ToothGrowth$len))
report(check_normality("len", ToothGrowth))
report(check_normality(ToothGrowth[["len"]]))
report(check_normality(runif(233)))
report(check_normality(rnorm(233)))
report(check_normality("ERROR"))
}
\seealso{
NCmisc::list.functions.in.file(filename = rstudioapi::getSourceEditorContext()$path)
}
