% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-time.R, R/compare.R
\name{compare.POSIXt}
\alias{compare}
\alias{compare.POSIXt}
\alias{compare.default}
\title{Provide human-readable comparison of two objects}
\usage{
\method{compare}{POSIXt}(x, y, tolerance = 0.001, ..., max_diffs = 9)

compare(x, y, ...)

\method{compare}{default}(x, y, ..., max_diffs = 9)
}
\arguments{
\item{x, y}{Objects to compare}

\item{tolerance}{Numerical tolerance: any differences smaller than this
value will be ignored. Default, 1e-5.}

\item{...}{Additional arguments used to control specifics of comparison}

\item{test.names}{If x and y are lists, whether to test the names of elements. Default, TRUE.}

\item{test.rowname}{If x and y are data frames, whether to test the row names. Default, FALSE.}

\item{test.colname}{If x and y are data frames, whether to test the column names. Default, TRUE.}
}
\description{
The parameters `test.rowname` and `test.colname` only works for data frames. While
the `test.names` parameter only works for list.
}
\details{
`compare` is similar to [base::all.equal()], but shows
you examples of where the failures occured.
Support data types: integer, numeric, logical, character, factor, list, matrix, data.frame
}

