% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{effsize_translate}
\alias{effsize_translate}
\title{Translate an effect size to another}
\usage{
effsize_translate(val, from, to, N = FALSE)
}
\arguments{
\item{val}{The current effect size value.}

\item{from}{The current effect size.}

\item{to}{The desired effect size.}

\item{N}{The total number of observations.}
}
\value{
Numeric translated effect size.
}
\description{
Transformation is based on psychometrica.
https://www.psychometrica.de/effektstaerke.html#transform
Accepted effect sizes:
\itemize{
\item from c(d, r, eta, f, chi, z)
\item to c(d, r, eta, f, chi, z)
}

'from'/'to' chi and z require N
}
\examples{
effsize_translate(-0.6761231, "z", "chi", 2044)
effsize_translate(-0.01495496, "r", "f")
}
