% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-handler.R
\name{error-handler}
\alias{addPostMsg}
\alias{addPreMsg}
\alias{error-handler}
\alias{registerPostMsg}
\alias{registerPreMsg}
\title{Self-defined error messages}
\usage{
registerPreMsg(...)

registerPostMsg(...)

addPreMsg(...)

addPostMsg(...)
}
\description{
Once there are unexpectations, add error messages in the output.
}
\details{
The arguments are exactly the same with the \code{\link{sprintf}} function.

\code{registerPreMsg} and \code{registerPostMsg} set up the customized messages
befor and after the error messages. If you set up multiple times, the last one will
overwritte all the previous settings.

\code{registerPreMsg} and \code{registerPostMsg} are doing append things instead of
overwritte.
}

