% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-to-yaml.R
\name{examples_to_yaml}
\alias{examples_to_yaml}
\title{examples_to_yaml}
\usage{
examples_to_yaml(
  package = NULL,
  functions = NULL,
  exclude = NULL,
  quiet = FALSE
)
}
\arguments{
\item{package}{Name of package, as either
\enumerate{
\item Path to local package source
\item Name of installed package
\item Full path to location of installed package if not on \link{.libPaths}, or
\item Default which presumes current directory is within package to be
tested.
}}

\item{functions}{If specified, names of functions from which examples are to
be obtained.}

\item{exclude}{Names of functions to exclude from 'yaml' template}

\item{quiet}{If 'FALSE', provide printed output on screen.}
}
\description{
Convert examples for a specified package, optionally restricted to one or
more specified functions, to a list of 'autotest' 'yaml' objects to use to
automatically test package.
}
\seealso{
Other yaml:
\code{\link[=at_yaml_template]{at_yaml_template()}},
\code{\link[=autotest_yaml]{autotest_yaml()}}
}
\concept{yaml}
