% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-expections.R
\name{expect_autotest_notes}
\alias{expect_autotest_notes}
\title{expect_autotest_notes}
\usage{
expect_autotest_notes(object)
}
\arguments{
\item{object}{An \code{autotest} object to be tested}
}
\value{
(invisibly) The same object
}
\description{
Expect \code{test_data} param of \code{autotest_package} to have additional \code{note}
column explaining why tests have been switched off.
}
\examples{
\donttest{
x <- autotest_package (package = "stats", functions = "cov", test = TRUE)
testthat::expect_success (expect_autotest_notes (x))
}
}
\seealso{
Other expectations:
\code{\link[=expect_autotest_no_err]{expect_autotest_no_err()}},
\code{\link[=expect_autotest_no_testdata]{expect_autotest_no_testdata()}},
\code{\link[=expect_autotest_no_warn]{expect_autotest_no_warn()}},
\code{\link[=expect_autotest_testdata]{expect_autotest_testdata()}}
}
\concept{expectations}
