% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structure.R
\name{force_structure}
\alias{force_structure}
\title{Call corresponding function of a \code{syscalls} pairlist}
\usage{
force_structure(syscalls, strctr)
}
\arguments{
\item{syscalls}{Function call from executing function as pairlist.}

\item{strctr}{Character vector, how to format data \code{c("numeric", "factor")}.}
}
\value{
Returns an evaluated function call.
}
\description{
The function receives a \code{syscalls} pairlist from the previous function call,
i.e. test_correl with a vector how the x and/or y variables shall be treated.
The original call is replaced with the appropriate \code{apply_structure_}-func-
tion. Structure information "numeric" or "factor" is appended to the call.
}
