% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_normality_table.R
\name{get_result}
\alias{get_result}
\title{Get the result of a normality test}
\usage{
get_result(test, data, expected, alpha = 0.05, alphacc = 0.3)
}
\arguments{
\item{test}{Shorthand of the test to run.}

\item{data}{A numeric vector of data values. Missing values are allowed.}

\item{expected}{Boolean of expected normality (TRUE|FALSE).}

\item{alpha}{Maximum accepted p-value.}

\item{alphacc}{Maximum accepted contingency coefficient for \code{jc} and \code{rc}.}
}
\value{
Integer 0|1.
}
\description{
Returns 1 if the test has confirmed the expected result.
Returns 0 if the test did not confirm the expected result.
}
\examples{
get_result("sf", rnorm(233), TRUE)
get_result("rc", rbeta(233, 10, 2), TRUE)
}
