% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{headline}
\alias{headline}
\title{Builds a headline (h1 - h4) for reporting}
\usage{
headline(text = FALSE, h = 1)
}
\arguments{
\item{text}{The text of the headline.}

\item{h}{The type/hierarchy of headline \code{c(1:4)}.}
}
\value{
returns a formatted headline.
}
\description{
Builds a headline (h1 - h4) for reporting
}
\examples{
headline("Hello World Headline", 1)
}
