% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_structure.R
\name{modify_factors}
\alias{modify_factors}
\title{Modifies factors which start with numbers or contain special characters}
\usage{
modify_factors(var, strctr)
}
\arguments{
\item{var}{The \code{v} from \code{apply_structure_}-functions.}

\item{strctr}{String of required structure.}
}
\value{
Return structure-formatted and modified data.
}
\description{
Modifies factors which start with numbers or contain special characters
}
\examples{
modify_factors(esoph$agegp, "factor")
modify_factors(ToothGrowth$supp, "factor")
}
