% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_table.R
\docType{data}
\name{normality_table}
\alias{normality_table}
\title{Table of successful normality tests}
\format{
\subsection{\code{normality_table}}{

A data frame with 25 rows and 11 columns:
\describe{
\item{n}{Sample size}
\item{winll}{Lilliefors}
\item{winsw}{Shapiro-Wilk}
\item{winpt}{Pearson with adjustment}
\item{winpf}{Pearson without adjustment}
\item{winad}{Anderson-Darling}
\item{wincv}{Cramer-von Mises}
\item{winsf}{Shapiro-Francia}
\item{winjb}{Jarque Bera}
\item{winjc}{Jarque Bera with contingency coefficient}
\item{max}{Maximum winner}
...
}
}
}
\source{
Function to build normality table
}
\usage{
normality_table
}
\description{
Table of successful normality tests
}
\keyword{datasets}
