% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{parse_exercise}
\alias{parse_exercise}
\title{Parse a Rmd file to an exercise}
\usage{
parse_exercise(input = "")
}
\arguments{
\item{input}{Input file (Rmd)}
}
\value{
A list, contains all the elements of exercise
}
\description{
Parse a Rmd file to an exercise
}
\examples{
path = system.file('rmarkdown/templates/skeleton/skeleton.Rmd', package = 'autotest')
parse_exercise(path)
}
