% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosLR_v1A_103020.R
\name{PosLR}
\alias{PosLR}
\title{PosLR}
\usage{
PosLR(
  object,
  sens.min = 0.5,
  spec.min = 0.5,
  top.n.genes = 10,
  alpha = 0.01,
  slot = "data",
  assay = "SCT",
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{sens.min}{Minimum sensitivity threshold. Default is 0.50}

\item{spec.min}{Minimum specificity threshold. Default is 0.50}

\item{top.n.genes}{Number of genes returned for each cluster. Default is 10}

\item{alpha}{Significance level used for calculating CI. Default is 0.01}

\item{slot}{Slot to pull data from. Default is "data"}

\item{assay}{Assay to use in differential expression testing. Default is "SCT"}
}
\value{
data.frame containing the top.n.genes most specific to each cluster
}
\description{
Finds differentially expressed genes for each cluster within a Seurat object
}
\details{
Returns a table containing genes most specific to each cluster, including sensitivity, specificity, CI and LR for each gene
}
\examples{
#Find top 5 DE genes for each cluster
PosLR(object = pbmc small, top.n.genes = 5)
}
