% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{print_htest}
\alias{print_htest}
\title{Pretty print a test result of S3 class 'htest'}
\usage{
print_htest(x)
}
\arguments{
\item{x}{The test result formatted as list.}
}
\value{
Returns a formatted list.
}
\description{
Function is based on package 'EnvStats'.
https://rdrr.io/cran/EnvStats/src/R/print.htestEnvStats.R
}
\examples{
print_htest(test_correl("len", "dose", ToothGrowth)$test[[1]])
print_htest(test_ttest("len", "supp", ToothGrowth)$test[[1]])
}
\seealso{
NCmisc::list.functions.in.file(filename = rstudioapi::getSourceEditorContext()$path)
}
