% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pstars}
\alias{pstars}
\title{Print stars of p-value}
\usage{
pstars(p, ls = FALSE, ts = FALSE)
}
\arguments{
\item{p}{The p-value of a significance test.}

\item{ls}{Leading space: adds space before stars.}

\item{ts}{Trailing space: adds space after stars.}
}
\value{
String.
}
\description{
Print stars of p-value
}
\examples{
pstars(.005)
}
\seealso{
NCmisc::list.functions.in.file(filename = rstudioapi::getSourceEditorContext()$path)
}
