% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{report}
\alias{report}
\title{Executes a string-saved call}
\usage{
report(x, elm = "report")
}
\arguments{
\item{x}{Stringed call, list parameter.}

\item{elm}{List element with stringified call to be executed.}
}
\value{
Runs the stringified function.
}
\description{
Executes a string-saved call
}
\examples{
report(test_correl("len", "dose", ToothGrowth, fast = FALSE), elm = "plot")
report(test_normality(rnorm(233), fast = FALSE), elm = "plot")
}
