% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{resultcol}
\alias{resultcol}
\title{Print the indicator of a test in green or red}
\usage{
resultcol(var, test)
}
\arguments{
\item{var}{The variable to be \code{TRUE} or \code{FALSE}.}

\item{test}{Shorthand \code{s}for significance, \code{n} for normality.}
}
\value{
returns a formatted string.
}
\description{
Print the indicator of a test in green or red
}
\examples{
cat(resultcol(TRUE, "s"))
cat(resultcol(FALSE, "n"))
}
