% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-case.R
\name{runTest}
\alias{runTest}
\title{Run test cases.}
\usage{
runTest(...)
}
\value{
A string of the format "AutoTest cases passed: {number of test cases}"
}
\description{
The \code{runTest()} function should be used to run test cases you defined
and it will generate output message if all tests have been passed.
}
