% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skip.R
\name{skip}
\alias{skip}
\alias{skip_if_not}
\alias{skip_if_not_installed}
\alias{skip_on_appveyor}
\alias{skip_on_bioc}
\alias{skip_on_cran}
\alias{skip_on_os}
\alias{skip_on_travis}
\title{Skip a test.}
\usage{
skip(message)

skip_if_not(condition, message = deparse(substitute(condition)))

skip_if_not_installed(pkg, minimum_version = NULL)

skip_on_cran()

skip_on_os(os)

skip_on_travis()

skip_on_appveyor()

skip_on_bioc()
}
\arguments{
\item{message}{A message describing why the test was skipped.}

\item{condition}{Boolean condition to check. If not `TRUE`, will
skip the test.}

\item{pkg}{Name of package to check for}

\item{minimum_version}{Minimum required version for the package}

\item{os}{Character vector of system names. Supported values are
`"windows"`, `"mac"`, `"linux"` and `"solaris"`.}
}
\description{
This function allows you to skip a test if it's not currently available.
This will produce an informative message, but will not cause the test
suite to fail.
}
\section{Helpers}{

`skip_if_not()` works like [stopifnot()], generating
a message automatically based on the first argument.

`skip_on_cran()` skips tests on CRAN, using the `NOT_CRAN`
environment variable set by devtools.

`skip_on_travis()` skips tests on travis by inspecting the
`TRAVIS` environment variable.

`skip_on_appveyor()` skips tests on appveyor by inspecting the
`APPVEYOR` environment variable.

`skip_on_bioc()` skips tests on Bioconductor by inspecting the
`BBS_HOME` environment variable.

`skip_if_not_installed()` skips a tests if a package is not installed
or cannot be loaded
(useful for suggested packages).
It loads the package as a side effect, because the package is likely to be
used anyway.
}
\examples{
if (FALSE) skip("No internet connection")
}

