% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_anova.R
\name{test_anova}
\alias{test_anova}
\title{Perform an ANOVA}
\usage{
test_anova(x, y, data = "", alpha = 0.05)
}
\arguments{
\item{x}{Name of column; dependent variable; metric values.}

\item{y}{Name of grouping column; independent variable; categorical.}

\item{data}{An optional data frame including \code{x} and \code{y} columns.}

\item{alpha}{Maximum accepted p-value.}
}
\value{
Returns a list of all results.
}
\description{
Perform an ANOVA
}
\examples{
test_anova("Sepal.Length", "Species", iris)
test_anova("Sepal.Width", "Species", iris)
test_anova("weight", "feed", chickwts)
}
