% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_anova.R
\name{test_anova.report}
\alias{test_anova.report}
\title{Class to build a full report for test_anova}
\usage{
test_anova.report(object)
}
\arguments{
\item{object}{Object of test_anova function}
}
\value{
Returns a full test report with simple figures
}
\description{
Class to build a full report for test_anova
}
\examples{
report(test_anova("weight", "feed", chickwts))
report(test_anova("Sepal.Length", "Species", iris))
report(test_anova("Sepal.Width", "Species", iris))
}
\seealso{
NCmisc::list.functions.in.file(filename = rstudioapi::getSourceEditorContext()$path)
}
