% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-case.R
\name{test_case}
\alias{test_case}
\title{Create a test case.}
\usage{
test_case(code)
}
\arguments{
\item{code}{test code containing expectations}
}
\value{
A test_that function containing the test code
}
\description{
A test case is a function that tests a series of expectations about
small, self-contained set of functionality. Each test is contained in a
\link{context} and contains multiple expectations.
}
\examples{
case_1 <- test_case({
  expect_equal(sin(pi / 4), 1 / sqrt(2))
  expect_equal(cos(pi / 4), 1 / sqrt(2))
  expect_equal(tan(pi / 4), 1)
})
case_1()
# Failing test:
\dontrun{
case_2 <- test_case(expect_equal(sin(pi / 4), 1))
case_2()
}
}
