% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_correl.R
\name{test_correl}
\alias{test_correl}
\title{Perform a bivariate correlation}
\usage{
test_correl(x, y, data = "", alternative = "two.sided", alpha = 0.05)
}
\arguments{
\item{x}{A numeric vector of data values.}

\item{y}{A numeric vector of data values.}

\item{data}{An optional data frame including \code{x} and \code{y} columns.}

\item{alternative}{Directed hypotheses; \code{greater} means positive correlation.}

\item{alpha}{Maximum accepted p-value.}
}
\value{
Returns a list of all results.
}
\description{
Perform a bivariate correlation
}
\examples{
test_correl("len", "dose", ToothGrowth)
test_correl(ToothGrowth$len, ToothGrowth$dose)
test_correl(ToothGrowth[["len"]], ToothGrowth[["dose"]])
test_correl(rnorm(2330), rnorm(2330))
}
