% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_correl.R
\name{test_correl.report}
\alias{test_correl.report}
\title{Class to build a full report for test_correl}
\usage{
test_correl.report(object)
}
\arguments{
\item{object}{Object of test_correl function}
}
\value{
Returns a full test report with simple figures
}
\description{
Class to build a full report for test_correl
}
\examples{
report(test_correl("len", "dose", ToothGrowth))
report(test_correl(ToothGrowth$len, ToothGrowth$dose))
report(test_correl(ToothGrowth[["len"]], ToothGrowth[["dose"]]))
report(test_correl(rnorm(2330), rnorm(2330)))
}
\seealso{
NCmisc::list.functions.in.file(filename = rstudioapi::getSourceEditorContext()$path)
}
