% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_crosstabs.R
\name{test_crosstabs}
\alias{test_crosstabs}
\title{Perform a crosstabs analysis}
\usage{
test_crosstabs(x, y, data = "", alternative = "two.sided", alpha = 0.05)
}
\arguments{
\item{x}{A numeric vector or a factor.}

\item{y}{A numeric vector or a factor of the same length as \code{x}.}

\item{data}{An optional data frame including \code{x} and \code{y} columns.}

\item{alternative}{Directed hypotheses; greater means \verb{group 1 > group 2} (ignored for df > 1).}

\item{alpha}{Maximum accepted p-value.}
}
\value{
Returns a list of all results.
}
\description{
Perform a crosstabs analysis
}
\examples{
test_crosstabs("cyl", "vs", mtcars)
test_crosstabs(mtcars$cyl, mtcars$vs)
test_crosstabs(mtcars[["cyl"]], mtcars[["vs"]])
}
\seealso{
NCmisc::list.functions.in.file(filename = rstudioapi::getSourceEditorContext()$path)
}
