% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_crosstabs.R
\name{test_crosstabs.report}
\alias{test_crosstabs.report}
\title{Class to build a full report for test_crosstabs}
\usage{
test_crosstabs.report(object)
}
\arguments{
\item{object}{Object of test_crosstabs function.}
}
\value{
Returns a full test report with simple figures.
}
\description{
Class to build a full report for test_crosstabs
}
\examples{
report(test_crosstabs("AirBags", "Man.trans.avail", Cars93))
report(test_crosstabs(Cars93$AirBags, Cars93$Man.trans.avail))
report(test_crosstabs(Cars93[["AirBags"]], Cars93[["Man.trans.avail"]]))
report(test_crosstabs("Origin", "Man.trans.avail", Cars93, alternative = "less"))
report(test_crosstabs("Origin", "Man.trans.avail", Cars93, alternative = "greater"))
}
\seealso{
NCmisc::list.functions.in.file(filename = rstudioapi::getSourceEditorContext()$path)
}
