% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_means.R
\name{test_means}
\alias{test_means}
\alias{test_ttest}
\title{Perform a parametric T-/Welch-Test or a non-parametric Wilcoxon/MWU test}
\usage{
test_means(
  x,
  y,
  data = "",
  paired = FALSE,
  alternative = "two.sided",
  alpha = 0.05
)

test_ttest(
  x,
  y,
  data = "",
  paired = FALSE,
  alternative = "two.sided",
  alpha = 0.05
)
}
\arguments{
\item{x}{Name of column; dependent variable; metric values.}

\item{y}{Name of grouping column; independent variable; categorical.}

\item{data}{An optional data frame including \code{x} and \code{y} columns.}

\item{paired}{TRUE if dependent groups, else \code{FALSE}.}

\item{alternative}{Directed hypotheses; greater means \verb{group 1 > group 2}.}

\item{alpha}{Maximum accepted p-value.}
}
\value{
Returns a list of all results.
}
\description{
Perform a parametric T-/Welch-Test or a non-parametric Wilcoxon/MWU test

Perform a parametric T-/Welch-Test or a non-parametric Wilcoxon/MWU test
}
\section{\code{test_ttest}}{

For \code{test_ttest}, use \code{\link{test_means}}.
}

\examples{
test_means("len", "supp", ToothGrowth)
test_means(ToothGrowth$len, ToothGrowth$supp)
test_means(ToothGrowth[["len"]], ToothGrowth[["supp"]])
}
