% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_means.R
\name{test_means.report}
\alias{test_means.report}
\title{Class to build a full report for test_means}
\usage{
test_means.report(object)
}
\arguments{
\item{object}{Object of test_means function}
}
\value{
Returns a full test report with simple figures
}
\description{
Class to build a full report for test_means
}
\examples{
report(test_means("len", "supp", ToothGrowth))
report(test_means(ToothGrowth$len, ToothGrowth$supp))
report(test_means(ToothGrowth[["len"]], ToothGrowth[["supp"]]))
}
