% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_normality.R
\name{test_normality}
\alias{test_normality}
\title{Performs a check on normality}
\usage{
test_normality(x, data = "", alpha = 0.05)
}
\arguments{
\item{x}{A numeric vector of data values.}

\item{data}{An optional data frame including \code{x} column.}

\item{alpha}{Maximum accepted p-value for significance tests.}
}
\value{
Returns a list of all results.
}
\description{
Performs a check on normality
}
\examples{
test_normality(ToothGrowth$len)
test_normality("len", ToothGrowth)
test_normality(ToothGrowth[["len"]])
test_normality(runif(233))
test_normality(rnorm(233))
test_normality("ERROR")
}
\seealso{
NCmisc::list.functions.in.file(filename = rstudioapi::getSourceEditorContext()$path)
}
