% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_normality.R
\name{test_normality.report}
\alias{test_normality.report}
\title{Class to build a full report for test_normality}
\usage{
test_normality.report(object)
}
\arguments{
\item{object}{Object of test_normality function}
}
\value{
Returns a full test report with simple figures
}
\description{
Class to build a full report for test_normality
}
\examples{
report(test_normality(ToothGrowth$len))
report(test_normality("len", ToothGrowth))
report(test_normality(ToothGrowth[["len"]]))
report(test_normality(runif(233)))
report(test_normality(rnorm(233)))
report(test_normality("ERROR"))
}
\seealso{
NCmisc::list.functions.in.file(filename = rstudioapi::getSourceEditorContext()$path)
}
