% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-package.R
\name{test_package}
\alias{test_check}
\alias{test_package}
\title{Run all tests in an installed package.}
\usage{
test_package(package, filter = NULL, reporter = default_reporter(), ...)

test_check(package, filter = NULL, reporter = "check", ...)
}
\arguments{
\item{package}{package name}

\item{filter}{If not `NULL`, only tests with file names matching this
regular expression will be executed.  Matching will take on the file
name after it has been stripped of `"test-"` and `".R"`.}

\item{reporter}{reporter to use}

\item{...}{Additional arguments passed to [grepl()] to control filtering.}
}
\value{
the results as a "autotest_results" (list)
}
\description{
Test are run in an environment that inherits from the package's namespace
environment, so that tests can access non-exported functions and variables.
Tests should be placed in `tests/autotest`. Use `test_check()` with
`R CMD check` and `test_package()` interactively at the console.
}
\section{R CMD check}{

Create `tests/autotest.R` that contains:

\preformatted{
library(autotest)
library(yourpackage)

test_check("yourpackage")
}
}
\examples{
\dontrun{test_package("autotest")}
}

