% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ttest.R
\name{test_ttest}
\alias{test_ttest}
\title{Perform a parametric T-/Welch-Test or a non-parametric Wilcoxon/MWU test}
\usage{
test_ttest(
  x,
  y,
  data = "",
  paired = FALSE,
  alternative = "two.sided",
  alpha = 0.05
)
}
\arguments{
\item{x}{Name of column; dependent variable; metric values.}

\item{y}{Name of grouping column; independent variable; categorical.}

\item{data}{An optional data frame including \code{x} and \code{y} columns.}

\item{paired}{TRUE if dependent groups, else \code{FALSE}.}

\item{alternative}{Directed hypotheses; greater means \verb{group 1 > group 2}.}

\item{alpha}{Maximum accepted p-value.}
}
\value{
Returns a list of all results.
}
\description{
Perform a parametric T-/Welch-Test or a non-parametric Wilcoxon/MWU test
}
\examples{
test_ttest("len", "supp", ToothGrowth)
test_ttest(ToothGrowth$len, ToothGrowth$supp)
test_ttest(ToothGrowth[["len"]], ToothGrowth[["supp"]])
}
\seealso{
NCmisc::list.functions.in.file(filename = rstudioapi::getSourceEditorContext()$path)
}
