% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ttest.R
\name{test_ttest.report}
\alias{test_ttest.report}
\title{Class to build a full report for test_ttest}
\usage{
test_ttest.report(object)
}
\arguments{
\item{object}{Object of test_ttest function}
}
\value{
Returns a full test report with simple figures
}
\description{
Class to build a full report for test_ttest
}
\examples{
report(test_ttest("len", "supp", ToothGrowth))
report(test_ttest(ToothGrowth$len, ToothGrowth$supp))
report(test_ttest(ToothGrowth[["len"]], ToothGrowth[["supp"]]))
}
\seealso{
NCmisc::list.functions.in.file(filename = rstudioapi::getSourceEditorContext()$path)
}
