% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-handler.R
\name{trace}
\alias{trace}
\alias{turnOffTraceback}
\alias{turnOnTraceback}
\title{Trun On/Off Tracking objects}
\usage{
turnOnTraceback()

turnOffTraceback()
}
\description{
Whether automatically detect the variables/expressions in \link{expect_equal}, \link{expect_true}, \link{expect_false} and etc.
Default it is TRUE.
}
\examples{
x = 1
\dontrun{
turnOnTraceback()
test_that('', {
  expect_equal(x, 2)
})
turnOffTraceback()   # turn off detecting, see the error message
test_that('', {
  expect_equal(x, 2)
})
}
}

