% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{at_yaml_template}
\alias{at_yaml_template}
\title{at_yaml_template}
\usage{
at_yaml_template()
}
\description{
Generate a 'yaml' template for an 'autotest' in the 'test' directory of
current package directory.
}
