% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotest-package.R
\docType{package}
\name{autotest-package}
\alias{autotest-package}
\alias{_PACKAGE}
\title{autotest: Automatic Package Testing}
\description{
Automatic testing of R packages via a simple YAML
    schema.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropenscilabs/autotest}
  \item Report bugs at \url{https://github.com/mpadge/ropenscilabs/autotest/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
