% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotest.R
\name{autotest_package}
\alias{autotest_package}
\title{autotest_package}
\usage{
autotest_package(
  package,
  functions = NULL,
  exclude = NULL,
  test = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{package}{Name of package to be 'autotested'}

\item{functions}{Optional character vector containing names of functions of
nominated package to be included in 'autotesting'.}

\item{exclude}{Optional character vector containing names of any functions of
nominated package to be excluded from 'autotesting'.}

\item{test}{If \code{FALSE}, return only descriptions of tests which would be run
with \code{test = TRUE}, without actually running them.}

\item{quiet}{If 'FALSE', provide printed output on screen.}
}
\description{
Automatically test an entire package by converting examples to \code{yaml} format
and submitting each to the \link{autotest} function.
}
