% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotest-functions.R
\name{autotest_package}
\alias{autotest_package}
\title{autotest_package}
\usage{
autotest_package(
  package = ".",
  functions = NULL,
  exclude = NULL,
  test = FALSE,
  test_data = NULL,
  quiet = FALSE
)
}
\arguments{
\item{package}{Name of package, as either
\enumerate{
\item Path to local package source
\item Name of installed package
\item Full path to location of installed package if not on \link{.libPaths}, or
\item Default which presumes current directory is within package to be
tested.
}}

\item{functions}{Optional character vector containing names of functions of
nominated package to be included in 'autotesting'.}

\item{exclude}{Optional character vector containing names of any functions of
nominated package to be excluded from 'autotesting'.}

\item{test}{If \code{FALSE}, return only descriptions of tests which would be run
with \code{test = TRUE}, without actually running them.}

\item{test_data}{Result returned from calling either \link{autotest_types} or
\link{autotest_package} with \code{test = FALSE} that contains a list of all tests
which would be conducted. These tests have an additional flag, \code{test}, which
defaults to \code{TRUE}. Setting any tests to \code{FALSE} will avoid running them when
\code{test = TRUE}.}

\item{quiet}{If 'FALSE', provide printed output on screen.}
}
\description{
Automatically test an entire package by converting examples to \code{yaml} format
and submitting each to the \link{autotest_yaml} function.
}
\seealso{
Other main_functions: 
\code{\link{autotest_types}()}
}
\concept{main_functions}
