% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotest.R
\name{autotest_types}
\alias{autotest_types}
\title{autotest_types}
\usage{
autotest_types()
}
\value{
An \code{autotest} object with each row listing one unique type of test
which can be applied to every parameter (of the appropriate class) of each
function.
}
\description{
List all types of 'autotests' currently implemented.
}
