% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autotest.R
\name{autotest_yaml}
\alias{autotest_yaml}
\title{autotest_yaml}
\usage{
autotest_yaml(yaml = NULL, filename = NULL, test = TRUE, quiet = FALSE)
}
\arguments{
\item{yaml}{A 'yaml' template as a character vector, either hand-coded or
potentially loaded via \link{readLines} function or similar. Should generally
be left at default of 'NULL', with template specified by 'filename'
parameter.}

\item{filename}{Name (potentially including path) of file containing 'yaml'
template. See \link{at_yaml_template} for details of template. Default uses
template generated by that function, and held in local './tests' directory.}

\item{test}{If \code{FALSE}, return only descriptions of tests which would be run
with \code{test = TRUE}, without actually running them.}

\item{quiet}{If 'FALSE', provide printed output on screen.}
}
\value{
An \code{autotest_pkg} object, derived from a \pkg{tibble}, detailing
instances of unexpected behaviour for every parameter of every function.
}
\description{
Automatically test inputs to functions specified in a 'yaml' template.
}
\examples{
\dontrun{
yaml_list <- examples_to_yaml (package = "stats", functions = "reshape")
res <- autotest_yaml (yaml = yaml_list)
}
}
