% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-handler.R
\name{error-handler}
\alias{error-handler}
\alias{registerPreMsg}
\alias{registerPostMsg}
\alias{addPreMsg}
\alias{addPostMsg}
\title{Self-defined error messages}
\usage{
registerPreMsg(...)

registerPostMsg(...)

addPreMsg(...)

addPostMsg(...)
}
\description{
Once there are unexpectations, add error messages in the output.
}
\details{
The arguments are exactly the same with the \code{\link{sprintf}} function.

\code{registerPreMsg} and \code{registerPostMsg} set up the customized messages
befor and after the error messages. If you set up multiple times, the last one will
overwritte all the previous settings.

\code{registerPreMsg} and \code{registerPostMsg} are doing append things instead of
overwritte.
}
\examples{
registerPreMsg('testing f(\%d)', 123) # just like sprintf('testing f(\%d)', 123)

f = function(x) ifelse(x<5, x+1, x)
# customize error messages
\dontrun{
test_that("", {
 for (i in 1:5){
   registerPreMsg("In testing f(\%d)", i)  # do not use `addPreMsg` here!
   expect_equal(f(i), i + 1)
 }
})
}
}
