% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-to-yaml.R
\name{examples_to_yaml}
\alias{examples_to_yaml}
\title{examples_to_yaml}
\usage{
examples_to_yaml(package = NULL)
}
\arguments{
\item{package}{Name of package for which 'yaml' is to be generated.}
}
\description{
Convert examples for a specified package to an 'autotest' 'yaml' to use to
automatically test package
}
