% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples-to-yaml.R
\name{examples_to_yaml}
\alias{examples_to_yaml}
\title{examples_to_yaml}
\usage{
examples_to_yaml(package = NULL, functions = NULL, exclude = NULL)
}
\arguments{
\item{package}{Name of package for which 'yaml' is to be generated.}

\item{functions}{If specified, names of functions from which examples are to
be obtained.}

\item{exclude}{Names of functions to exclude from 'yaml' template}
}
\description{
Convert examples for a specified package, optionally restricted to one or
more specified functions, to a list of 'autotest' 'yaml' objects to use to
automatically test package.
}
