% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-expections.R
\name{expect_autotest_no_err}
\alias{expect_autotest_no_err}
\title{expect_autotest_no_err}
\usage{
expect_autotest_no_err(object)
}
\arguments{
\item{object}{An \code{autotest} object to be tested}
}
\value{
(invisibly) The same object
}
\description{
Expect \code{autotest_package()} to be clear of errors
}
\seealso{
Other expectations: 
\code{\link{expect_autotest_no_testdata}()},
\code{\link{expect_autotest_no_warn}()},
\code{\link{expect_autotest_notes}()},
\code{\link{expect_autotest_testdata}()}
}
\concept{expectations}
