% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-expections.R
\name{expect_autotest_no_testdata}
\alias{expect_autotest_no_testdata}
\title{expect_autotest_no_testdata}
\usage{
expect_autotest_no_testdata(object = NULL)
}
\arguments{
\item{object}{Not used here, but required for \code{testthat} expectations}
}
\value{
(invisibly) The autotest object
}
\description{
Expect \code{autotest_package()} to be clear of errors with no tests switched off
}
\seealso{
Other expectations: 
\code{\link{expect_autotest_no_err}()},
\code{\link{expect_autotest_no_warn}()},
\code{\link{expect_autotest_notes}()},
\code{\link{expect_autotest_testdata}()}
}
\concept{expectations}
