% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-expections.R
\name{expect_autotest_no_testdata}
\alias{expect_autotest_no_testdata}
\title{expect_autotest_no_testdata}
\usage{
expect_autotest_no_testdata(object = NULL)
}
\arguments{
\item{object}{Not used here, but required for \code{testthat} expectations}
}
\value{
(invisibly) The autotest object
}
\description{
Expect \code{autotest_package()} to be clear of errors with no tests switched off
}
