% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-expections.R
\name{expect_autotest_notes}
\alias{expect_autotest_notes}
\title{expect_autotest_notes}
\usage{
expect_autotest_notes(object)
}
\arguments{
\item{object}{An \code{autotest} object to be tested}
}
\description{
Expect \code{test_data} param of \code{autotest_package} to have additional \code{note}
column explaining why tests have been switched off.
}
