% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-expections.R
\name{expect_autotest_testdata}
\alias{expect_autotest_testdata}
\title{expect_autotest_testdata}
\usage{
expect_autotest_testdata(object)
}
\arguments{
\item{object}{An \code{autotest_package} object with a \code{test} column flagging
tests which are not to be run on the local package.}
}
\value{
(invisibly) The autotest object
}
\description{
Expect \code{autotest_package()} to be clear of errors with some tests switched
off, and to have note column explaining why those tests are not run.
}
\examples{
test_data <- autotest_types (notest = "vector_to_list_col")
\dontrun{
# Called within a 'testthat' test file of the local package itself:
testthat::test_that ("autotest", {
    testthat::expect_success (expect_autotest_testdata (test_data))
})
}
}
\seealso{
Other expectations:
\code{\link[=expect_autotest_no_err]{expect_autotest_no_err()}},
\code{\link[=expect_autotest_no_testdata]{expect_autotest_no_testdata()}},
\code{\link[=expect_autotest_no_warn]{expect_autotest_no_warn()}},
\code{\link[=expect_autotest_notes]{expect_autotest_notes()}}
}
\concept{expectations}
