% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_table.R
\docType{data}
\name{normality_table}
\alias{normality_table}
\title{Table of successful normality tests}
\format{
\subsection{\code{normality_table}}{

A data frame with 26 rows and 15 columns:
\describe{
\item{n}{Sample size}
\item{ad}{Anderson-Darling}
\item{cv}{Cramer-von Mises}
\item{da}{D'Agostino}
\item{jb}{Jarque Bera (classic)}
\item{jc}{Jarque–Bera CC (classic)}
\item{ks}{Kolmogorov-Smirnov}
\item{ll}{Lilliefors}
\item{pf}{Pearson chi-square}
\item{pt}{Pearson chi-square (adjusted)}
\item{rb}{Jarque–Bera (robust)}
\item{rb}{Jarque–Bera CC (robust)}
\item{sf}{Shapiro-Francia}
\item{sw}{Shapiro-Wilk}
\item{max}{Maximum winner}
}
}
}
\source{
build_normality_table.R
}
\usage{
normality_table
}
\description{
Table of successful normality tests
}
\seealso{
\code{\link[=build_normality_table]{build_normality_table()}}
}
\keyword{datasets}
