% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{reportp}
\alias{reportp}
\title{Formats the p.value for reporting purposes}
\usage{
reportp(p)
}
\arguments{
\item{p}{The p.value.}
}
\value{
Returns the p-value as string (e.g., p < .001)
}
\description{
Formats the p.value for reporting purposes
}
\examples{
reportp(0.00012)
reportp(0.00123)
reportp(0.01234)
reportp(0.13456)
}
