% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_normality_table.R
\name{run_test}
\alias{run_test}
\title{Runs a normality test}
\usage{
run_test(test, data, alpha = 0.05, alphacc = 0.3)
}
\arguments{
\item{test}{Shorthand of the test to run.}

\item{data}{A numeric vector of data values. Missing values are allowed.}

\item{alpha}{Maximum accepted p-value.}

\item{alphacc}{Maximum accepted contingency coefficient for \code{jc} and \code{rc}.}
}
\value{
Boolean if normal.
}
\description{
Runs a normality test
}
