% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_crosstabs.R
\name{test_crosstabs.report}
\alias{test_crosstabs.report}
\title{Class to build a full report for test_crosstabs}
\usage{
test_crosstabs.report(object)
}
\arguments{
\item{object}{Object of test_crosstabs function.}
}
\value{
Returns a full test report with simple figures.
}
\description{
Class to build a full report for test_crosstabs
}
\examples{
report(test_crosstabs("cyl", "vs", mtcars))
report(test_crosstabs(mtcars$cyl, mtcars$vs))
report(test_crosstabs(mtcars[["cyl"]], mtcars[["vs"]]))
}
\seealso{
NCmisc::list.functions.in.file(filename = rstudioapi::getSourceEditorContext()$path)
}
