% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_normality.R
\name{test_normality}
\alias{test_normality}
\title{Performs a check on normality}
\usage{
test_normality(x, data = "", alpha = 0.05, alphacc = 0.3)
}
\arguments{
\item{x}{A numeric vector of data values.}

\item{data}{An optional data frame including \code{x} column.}

\item{alpha}{Maximum accepted p-value for significance tests.}

\item{alphacc}{Maximum accepted contingency coefficient for \code{jc} and \code{rc}.}
}
\value{
Returns a list of all results.
}
\description{
Performs a check on normality
}
\examples{
var <- cat(test_normality(ToothGrowth$len)$result)
test_normality("len", ToothGrowth)
test_normality(ToothGrowth[["len"]])
test_normality(runif(233))
test_normality(rnorm(89))
}
